Unified naming scheme
