+++ tests/test_articlehistory.py
@@ -1,15 +1,37 @@
-""" to be merged in with nob/1 """
-
+'''test for article history subclass'''
 import __init__
-from articlerevisions import ArticleRevisions
+import pytest
+from articlehistory import ArticleHistory, format_timestamp
 
+def test___init__():
+    '''tests initalization'''
+    art = ArticleHistory("Techno", "Rio65trio")
+    assert art.titles == "Techno"
+    assert art.user == "Rio65trio"
+    assert len(art.revisions) == 10
+    assert art.pageid == 23958411
+
+    art = ArticleHistory("Cat")
+    assert art.titles == "Cat"
+    assert art.user is None
+    assert len(art.revisions) != 0
+
+    art = ArticleHistory("Salsa", "CAPTAIN RAJU")
+    assert art.titles == "Salsa"
+    assert art.user == "CAPTAIN RAJU"
+    assert len(art.revisions) == 1
+
+    art = ArticleHistory("fdjaklfgd;jsa")
+    with pytest.raises(KeyError):
+        raise KeyError("Data not found")
+
+    art = ArticleHistory("")
+    with pytest.raises(KeyError):
+        raise KeyError("Data not found")
+        
 def test_format_timestamp():
     """ tests module level function for creating timestamp string from ints"""
-    t_revisions = ArticleRevisions("test")
+    t_revisions = ArticleHistory("test")
     # full datetime
-    assert t_revisions.format_timestamp(2008, 6, 7, 4, 1, 30) == '20080607040130'
+    assert format_timestamp(2008, 6, 7, 4, 1, 30) == '20080607040130'
     # partials to do
-    #assert a.format_timestamp(2008, 0, 7, 4, 1, 30) == '20080607040130'
-
-if __name__ == "__main__":
-    test_format_timestamp()
